move hickory-resolved dep to amaru-network
