+++ crates/amaru-network/Cargo.toml
@@ -17,6 +17,7 @@ categories.workspace = true
 
 [dependencies]
 bytes = { workspace = true, features = ["serde"] }
+hickory-resolver.workspace = true
 parking_lot.workspace = true
 socket2.workspace = true
 tokio = { workspace = true, features = ["net"] }
