+++ crates/amaru-network/src/lib.rs
@@ -13,3 +13,4 @@
 // limitations under the License.
 
 pub mod connection;
+pub mod resolve;
